fix: floating point is like 6k!!! added cmd for display
adding a proper floating point operation added something like 6 kb to
the resulting bin size, which is not acceptable, I'm pretty tight with
the graphics library! Also added a separate packet/command for
refreshing the display directly, and stopped redrawing the display for
each led update. This allows for a mixed experience, with smooth leds
with "hiccups" and a slow updating display.